Add admin module serving the bucket web pages
