+++ cmd/profilecli/static/fonts/bootstrap-icons.woff
