+++ cmd/profilecli/static/pyroscope-logo.png
