+++ cmd/profilecli/static/pyroscope-styles.css
