+++ pkg/operations/tool.blocks.detail.gohtml
@@ -7,10 +7,10 @@
 
     <title>Bucket Blocks Explorer: Tenant {{ .User }}, Block {{ .Block.ID }}</title>
 
-    <link rel="stylesheet" href="/tool/static/bootstrap-5.1.3.min.css">
-    <link rel="stylesheet" href="/tool/static/bootstrap-icons-1.8.1.css">
-    <link rel="stylesheet" href="/tool/static/pyroscope-styles.css">
-    <script src="/tool/static/bootstrap-5.1.3.bundle.min.js"></script>
+    <link rel="stylesheet" href="/static/bootstrap-5.1.3.min.css">
+    <link rel="stylesheet" href="/static/bootstrap-icons-1.8.1.css">
+    <link rel="stylesheet" href="/static/pyroscope-styles.css">
+    <script src="/static/bootstrap-5.1.3.bundle.min.js"></script>
 </head>
 <body class="d-flex flex-column h-100">
 <main class="flex-shrink-0">
@@ -20,14 +20,14 @@
                 <h3>Bucket Blocks Explorer: Block Details</h3>
             </div>
             <div class="col-12 col-sm-3 text-center text-sm-end mb-3 mb-sm-0">
-                <a href="/tenants">
-                    <img alt="Pyroscope logo" class="pyroscope-brand" src="/tool/static/pyroscope-logo.png">
+                <a href="/ops/object-store/tenants">
+                    <img alt="Pyroscope logo" class="pyroscope-brand" src="/static/pyroscope-logo.png">
                 </a>
             </div>
         </div>
         <div class="row my-3">
             <p>
-                <a href="/blocks/list?tenantId={{ .User }}">Back to tenant</a>
+                <a href="/ops/object-store/tenants/{{ .User }}/blocks">Back to tenant</a>
             </p>
             <table class="table">
                 <tr>
