+++ pkg/operations/tool.blocks.index.gohtml
@@ -7,10 +7,10 @@
 
     <title>Bucket Blocks Explorer: Grafana Pyroscope</title>
 
-    <link rel="stylesheet" href="/tool/static/bootstrap-5.1.3.min.css">
-    <link rel="stylesheet" href="/tool/static/bootstrap-icons-1.8.1.css">
-    <link rel="stylesheet" href="/tool/static/pyroscope-styles.css">
-    <script src="/tool/static/bootstrap-5.1.3.bundle.min.js"></script>
+    <link rel="stylesheet" href="/static/bootstrap-5.1.3.min.css">
+    <link rel="stylesheet" href="/static/bootstrap-icons-1.8.1.css">
+    <link rel="stylesheet" href="/static/pyroscope-styles.css">
+    <script src="/static/bootstrap-5.1.3.bundle.min.js"></script>
 </head>
 <body class="d-flex flex-column h-100">
 <main class="flex-shrink-0">
@@ -20,8 +20,8 @@
                 <h3>Bucket Blocks Explorer: Grafana Pyroscope</h3>
             </div>
             <div class="col-12 col-sm-3 text-center text-sm-end mb-3 mb-sm-0">
-                <a href="/blocks/index">
-                    <img alt="Pyroscope logo" class="pyroscope-brand" src="/tool/static/pyroscope-logo.png">
+                <a href="/ops/object-store/tenants">
+                    <img alt="Pyroscope logo" class="pyroscope-brand" src="/static/pyroscope-logo.png">
                 </a>
             </div>
         </div>
@@ -39,7 +39,7 @@
                     {{ range $i, $ := .Users }}
                         <tr>
                             <td class="align-middle font-monospace small">
-                                <a href='/blocks/list?tenantId={{ . }}'>{{ . }}</a>
+                                <a href='/ops/object-store/tenants/{{ . }}/blocks'>{{ . }}</a>
                             </td>
                         </tr>
                     {{ end }}
